last digit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -7,7 +7,7 @@
 int print_last_digit(int i)
 {
 	int ld;
-	unsigned int j;
+	long j;
 
 	if (i < 0)
 		i = -i;
